Moved ssi tests into syntax_tests/test_ssi.py.
